move quorum bits around
