+++ src/BrightChain.Engine/Quorum/Services/ConfidentialQuorum.cs
@@ -1,0 +1,6 @@
+namespace BrightChain.Engine.Quorum.Services;
+
+public class ConfidentialQuorum
+{
+    
+}
